splitting out the struct definitions to a seperate file
