+++ burrito_link/Makefile
@@ -6,7 +6,7 @@ CFLAGS=-Wall -Os -g -lws2_32
 all: burrito_link.exe
 
 # Create the CLI Executable
-burrito_link.exe: gw2mumbleudp.c
+burrito_link.exe: gw2mumbleudp.c linked_memory.h
 	$(WINECC) $< $(CFLAGS) -mconsole -o $@
 
 # Remove all targets and generated files
